first deploy

+++ src/controllers/group.controller.js
@@ -1,6 +1,7 @@
 const Project = require("../models/project.model");
 const Group = require("../models/group.model");
 const User = require("../models/user.model");
+const Technician = require("../models/technician.model");
 
 const create = async (req, res) => {
   try {
@@ -41,7 +42,7 @@ const list = async (req, res) => {
   try {
     const { userId } = req;
     const { projectId } = req.params;
- 
+
     const user = await User.findById(userId);
     if (!user) {
       throw new Error("Invalid user");
@@ -70,7 +71,7 @@ const show = async (req, res) => {
     if (!user) {
       throw new Error("Invalid user");
     }
-    const group = await Group.findById(groupId);    
+    const group = await Group.findById(groupId);
     if (!group) {
       throw new Error("Invalid group");
     }
@@ -91,7 +92,7 @@ const destroy = async (req, res) => {
   try {
     const { groupId } = req.params;
     const { userId } = req;
-    console.log(groupId)
+    console.log(groupId);
     const user = await User.findById(userId);
     if (!user) {
       throw new Error("Invalid user");
@@ -101,19 +102,28 @@ const destroy = async (req, res) => {
       throw new Error("Invalid group");
     }
     const project = await Project.findById(group.projectId);
-    console.log(project)
+
     if (project.user.toString() !== user._id.toString()) {
       throw new Error("project does not belong to this user");
     }
-    await Group.findByIdAndDelete(group._id);
-    await Project.updateOne(
-      { _id: project._id },
-      {
-        $pull: { groupsId: group._id },
-      }
-    );
 
-    res.status(200).json({ ok: true, message: "group deleted", data: group });
+    if (group.techniciansId.length === 0) {
+      await Group.findByIdAndDelete(group._id);
+      await Project.updateOne(
+        { _id: project._id },
+        {
+          $pull: { groupsId: group._id },
+        }
+      );
+
+      res.status(200).json({ ok: true, message: "group deleted", data: group });
+    } else {
+      res.status(400).json({
+        ok: false,
+        message:
+          "the group could not be deleted with the technicians included in its array",
+      });
+    }
   } catch (err) {
     console.log(err);
     res
@@ -122,9 +132,196 @@ const destroy = async (req, res) => {
   }
 };
 
+const addOrRemoveTechnician = async (req, res) => {
+  const groupId = req.params.groupId;
+  const technicianId = req.params.technicianId;
+  const { userId } = req;
+
+  try {
+    const group = await Group.findById(groupId);
+    if (!group) {
+      throw new Error("Invalid group");
+    }
+
+    const user = await User.findById(userId);
+    if (!user) {
+      throw new Error("Invalid user");
+    }
+
+    const technician = await Technician.findById(technicianId);
+    if (!technician) {
+      throw new Error("Invalid technician");
+    }
+
+    const findTechnical = group.techniciansId.includes(technician._id);
+
+    if (!findTechnical) {
+      await Group.updateOne(
+        { _id: group._id },
+        {
+          $push: { techniciansId: technician._id },
+        }
+      );
+
+      const groupUpdate = await Group.findById(group._id);
+      return res.status(200).json({
+        ok: true,
+        msg: "technical successfully added",
+        group: groupUpdate,
+      });
+    } else {
+      await Group.updateOne(
+        { _id: group._id },
+        {
+          $pull: {
+            techniciansId: technician._id,
+          },
+        }
+      );
+      const groupUpdate = await Group.findById(group._id);
+      return res.status(200).json({
+        ok: true,
+        msg: "technical successfully removed",
+        group: groupUpdate,
+      });
+    }
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+const addOrRemoveTechnicianInProject = async (req, res) => {
+  const groupId = req.params.groupId;
+  const technicianId = req.params.technicianId;
+  const { userId } = req;
+
+  try {
+    const group = await Group.findById(groupId);
+    if (!group) {
+      throw new Error("Invalid group");
+    }
+    const project = await Project.findById(group.projectId);
+    if (!project) {
+      throw new Error("Invalid project");
+    }
+
+    const user = await User.findById(userId);
+    if (!user) {
+      throw new Error("Invalid user");
+    }
+
+    const technician = await Technician.findById(technicianId);
+    if (!technician) {
+      throw new Error("Invalid technician");
+    }
+
+    const findTechnical = group.techniciansId.includes(technician._id);
+
+    if (!findTechnical) {
+      await Group.updateOne(
+        { _id: group._id },
+        {
+          $push: { techniciansId: technician._id },
+        }
+      );
+      await Project.updateOne(
+        { _id: project._id },
+        {
+          $push: { techniciansId: technician._id },
+        }
+      );
+      await Technician.updateOne(
+        { _id: technician._id },
+        {
+          // $push: { techniciansId: technician._id },
+          projectId: project._id,
+        }
+      );
+
+      const projectUpdate = await Project.findById(project._id);
+      return res.status(200).json({
+        ok: true,
+        msg: "technical successfully added",
+        project: projectUpdate,
+      });
+    } else {
+      await Group.updateOne(
+        { _id: group._id },
+        {
+          $pull: {
+            techniciansId: technician._id,
+          },
+        }
+      );
+      await Project.updateOne(
+        { _id: project._id },
+        {
+          $pull: { techniciansId: technician._id },
+        }
+      );
+      await Technician.updateOne(
+        { _id: technician._id },
+        {
+          // $push: { techniciansId: technician._id },
+          projectId: null,
+        }
+      );
+      const projectUpdate = await Project.findById(project._id);
+      return res.status(200).json({
+        ok: true,
+        msg: "technical successfully removed",
+        project: projectUpdate,
+      });
+    }
+  } catch (err) {
+    console.log(err);
+  }
+};
+
+const update = async (req, res) => {
+  try {
+    const { groupId } = req.params;
+    const { userId } = req;
+    console.log(groupId);
+    const user = await User.findById(userId);
+    if (!user) {
+      throw new Error("Invalid user");
+    }
+    const group = await Group.findById(groupId);
+    if (!group) {
+      throw new Error("Invalid group");
+    }
+    const project = await Project.findById(group.projectId);
+
+    if (project.user.toString() !== user._id.toString()) {
+      throw new Error("project does not belong to this user");
+    }
+
+    const groupUpdate = await Group.findByIdAndUpdate(group._id, req.body, {
+      new: true,
+      runValidators: true,
+      context: "query",
+    });
+    res.status(200).json({
+      ok: true,
+      message: "group updated",
+      data: groupUpdate,
+    });
+  } catch (err) {
+    res.status(500).json({
+      ok: false,
+      message: "group could not be update",
+      data: err.message,
+    });
+  }
+};
+
 module.exports = {
   create,
   list,
   show,
   destroy,
+  addOrRemoveTechnician,
+  addOrRemoveTechnicianInProject,
+  update,
 };
